+++ cx-portal/src/components/pages/MyAccount/index.tsx
@@ -46,7 +46,7 @@ export default function MyAccount() {
   const dispatch = useDispatch()
 
   const handleDeleteUser = () =>
-    dispatch(show(OVERLAYS.CONFIRM_USER_ACTION, data?.companyUserId, 'admin'))
+    dispatch(show(OVERLAYS.CONFIRM_USER_ACTION, data?.companyUserId, 'ownUser'))
 
   return (
     <main className="my-account">