+++ cx-portal/src/types/Patterns.test.ts
@@ -27,6 +27,8 @@ import {
   isMail,
   isURL,
   isUUID,
+  isCName,
+  isCountryCode,
 } from './Patterns'
 
 const TESTDATA = {
@@ -137,6 +139,14 @@ const TESTDATA = {
     ],
     invalid: ['', '   ', ' Company ', '"!§$%&/'],
   },
+  CNAMES: {
+    valid: ['word', 'Some word', 'Some@word'],
+    invalid: ['some.word', 'w', '', 'some?@#$word'],
+  },
+  COUNTRY: {
+    valid: ['DE'],
+    invalid: ['', 'D', 'de', 'some string'],
+  },
 }
 
 describe('Input Pattern Tests', () => {
@@ -173,6 +183,20 @@ describe('Input Pattern Tests', () => {
     )
     TESTDATA.COMPANY_NAME.invalid.forEach((expr) =>
       expect(isCompanyName(expr)).toBe(false)
+    )
+  })
+
+  it('validates connector Names', () => {
+    TESTDATA.CNAMES.valid.forEach((expr) => expect(isCName(expr)).toBe(true))
+    TESTDATA.CNAMES.invalid.forEach((expr) => expect(isCName(expr)).toBe(false))
+  })
+
+  it('validates Country code', () => {
+    TESTDATA.COUNTRY.valid.forEach((expr) =>
+      expect(isCountryCode(expr)).toBe(true)
+    )
+    TESTDATA.COUNTRY.invalid.forEach((expr) =>
+      expect(isCountryCode(expr)).toBe(false)
     )
   })
 })