+++ CHANGELOG.md
@@ -5,8 +5,15 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
 
 ### Unreleased
 
-* ...
-* ...
+* Register connector
+    * Enable "Company as a Service" option. Modal updates
+    * Validation updates. FormFields added for "Company as a Service".
+    * Create & Delete API integration with flow
+* Bugfix - User Invite - New User Role preselected fix
+* Bugfix - App User Management - Confirmation and User Role Changes
+* Shared Components
+   * More versatile image component
+* Feature - User Management - fetch roles (api refactoring)
 
 
 ## 0.7.0